Publish config-server structs as proto
Currently, config-server serializes the old ChainedServerInfo from this
repository, while flashlight deserializes it with a random proto definition from
lantern-cloud. This will enable us to switch out flashlight's proto definition
to point to this one, and update config-server to point to this, so that server
and client agree on the message format.